feat(exp): rename `*utils` package to `*util`
