+++ hcloud/exp/actionutil/actions_test.go
@@ -1,4 +1,4 @@
-package actionutils
+package actionutil
 
 import (
 	"testing"
