+++ hcloud/exp/kit/envutil/env.go
@@ -1,4 +1,4 @@
-package envutils
+package envutil
 
 import (
 	"fmt"
