+++ hcloud/exp/kit/envutil/env_test.go
@@ -1,4 +1,4 @@
-package envutils
+package envutil
 
 import (
 	"os"
