+++ hcloud/exp/kit/sshutil/ssh_key.go
@@ -1,4 +1,4 @@
-package sshutils
+package sshutil
 
 import (
 	"crypto"
