+++ hcloud/exp/kit/sshutil/ssh_key_test.go
@@ -1,4 +1,4 @@
-package sshutils
+package sshutil
 
 import (
 	"strings"
