+++ hcloud/exp/labelutil/selector_test.go
@@ -1,4 +1,4 @@
-package labelutils
+package labelutil
 
 import "testing"
