+++ hcloud/exp/mockutil/http_test.go
@@ -1,4 +1,4 @@
-package mockutils
+package mockutil
 
 import (
 	"fmt"
